test: split the delete cases out of the lifecycle suite
Adding the three-outcome coverage pushed api-handlers-lifecycle.test.ts to
539 lines, over the 500-line cap. The delete cases move to their own file;
same tests, same assertions, no behavior change.
